+++ src/main/java/org/sopt/member/application/service/MemberDeleteService.java
@@ -5,18 +5,18 @@ import static org.sopt.member.domain.exception.MemberErrorCode.MEMBER_NOT_FOUND;
 import java.util.NoSuchElementException;
 import org.sopt.member.application.port.in.MemberDeleteUseCase;
 import org.sopt.member.domain.exception.MemberException;
-import org.sopt.member.domain.port.out.MemberRepositoryPort;
+import org.sopt.member.domain.port.out.MemberRepository;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemberDeleteService implements MemberDeleteUseCase {
 
-    private final MemberRepositoryPort memberRepository;
+    private final MemberRepository memberRepository;
 
     public MemberDeleteService(
-        MemberRepositoryPort memberRepositoryPort
+        MemberRepository memberRepository
     ) {
-        this.memberRepository = memberRepositoryPort;
+        this.memberRepository = memberRepository;
     }
 
     @Override
