+++ src/main/java/org/sopt/member/application/service/MemberFindOneService.java
@@ -7,14 +7,14 @@ import org.sopt.member.application.dto.result.MemberFindOneResult;
 import org.sopt.member.application.port.in.MemberFindOneUseCase;
 import org.sopt.member.domain.exception.MemberException;
 import org.sopt.member.domain.model.Member;
-import org.sopt.member.domain.port.out.MemberRepositoryPort;
+import org.sopt.member.domain.port.out.MemberRepository;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
 public class MemberFindOneService implements MemberFindOneUseCase {
 
-    private final MemberRepositoryPort memberRepository;
+    private final MemberRepository memberRepository;
 
     @Override
     public MemberFindOneResult findOne(Long id) {
