+++ src/main/java/org/sopt/member/application/service/MemberJoinService.java
@@ -8,18 +8,18 @@ import org.sopt.member.application.dto.result.MemberJoinResult;
 import org.sopt.member.application.port.in.MemberJoinUseCase;
 import org.sopt.member.domain.exception.MemberException;
 import org.sopt.member.domain.model.Member;
-import org.sopt.member.domain.port.out.MemberRepositoryPort;
+import org.sopt.member.domain.port.out.MemberRepository;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemberJoinService implements MemberJoinUseCase {
 
-    private final MemberRepositoryPort memberRepository;
+    private final MemberRepository memberRepository;
 
     public MemberJoinService(
-        MemberRepositoryPort memberRepositoryPort
+        MemberRepository memberRepository
     ) {
-        this.memberRepository = memberRepositoryPort;
+        this.memberRepository = memberRepository;
     }
 
     @Override
