+++ src/main/java/org/sopt/member/infrastructure/jpa/MemberJpaRepository.java
@@ -1,9 +1,9 @@
-package org.sopt.member.domain.port.out;
+package org.sopt.member.infrastructure.jpa;
 
 import org.sopt.member.domain.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepositoryPort extends JpaRepository<Member, Long> {
+public interface MemberJpaRepository extends JpaRepository<Member, Long> {
 
     boolean existsByEmail(String email);
